testfiles/requirements.txt
pyserial
tqdm
numpy
pandas